Exception Handler, obsługa wyjątków controllera

+++ src/main/java/com/songify/song/controller/SongRestController.java
@@ -1,5 +1,10 @@
-package com.songify.song;
+package com.songify.song.controller;
 
+import com.songify.song.dto.DeleteSongResponseDto;
+import com.songify.song.dto.SingleSongResponseDto;
+import com.songify.song.dto.SongRequestDto;
+import com.songify.song.dto.SongResponseDto;
+import com.songify.song.error.SongNotFoundException;
 import jakarta.validation.Valid;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.http.HttpStatus;
@@ -50,9 +55,11 @@ public class SongRestController {
         return ResponseEntity.ok(new SingleSongResponseDto(songName));
     }
     @DeleteMapping("/song/{id}")
-    public ResponseEntity<String> deleteSongByIdUsingPathVariable(@PathVariable  Integer id){
+    public ResponseEntity<DeleteSongResponseDto> deleteSongByIdUsingPathVariable(@PathVariable  Integer id){
+        if(!database.containsKey(id)){
+            throw new SongNotFoundException("Song with id " + id + " not found");
+        }
         database.remove(id);
-        return ResponseEntity.ok("You deleted song with id: " + id);
+        return ResponseEntity.ok(new DeleteSongResponseDto("You deleted song by id: " + id, HttpStatus.OK));
     }
-
 }
